Add a function to convert bpch diagnostic names to netCDF names
Added function convert_bpch_names_to_netcdf_names to the
benchmark_1mon_prototype.ipynb notebook.  This contains a large
dictionary with the bpch diagnostic names and their new netCDF
diagnostic equivalents.  Use this to change the variable names in
an xarray Dataset object that was opened with xbpch.open_bpchdataset.

NOTE: I couldn't really test this update because the notebook is
pegged to Lizzie's Mac.  Some minor editing might be necessary but
this should work.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>